last commit beofre merge , drag works but needs more work so i will leave the package gave DELETE route for boards a success and that removed some of the errors

+++ app/api/boards/route.js
@@ -1,4 +1,3 @@
-
 import { NextResponse } from "next/server";
 import { auth } from "@/auth";
 
@@ -6,13 +5,10 @@ import User from "@/models/User";
 import Board from "@/models/Board";
 import { connectMongo } from "@/lib/mongoose";
 
-
-
 export async function POST(req) {
   try {
     const body = await req.json();
     if (!body.name) {
-
       return NextResponse.json(
         { error: "Board Name is Required" },
         { status: 400 }
@@ -27,7 +23,6 @@ export async function POST(req) {
 
     await connectMongo();
 
-
     const user = await User.findById(session.user.id);
 
     const board = await Board.create({
@@ -38,7 +33,6 @@ export async function POST(req) {
     user.boards.push(board._id);
 
     await user.save();
-
 
     return NextResponse.json(board);
   } catch (error) {
@@ -56,13 +50,11 @@ export async function GET() {
 
     await connectMongo();
 
-
     const user = await User.findById(session.user.id).populate("boards");
 
     if (!user) {
       return NextResponse.json({ error: "User not found" }, { status: 404 });
     }
-
 
     return NextResponse.json(user.boards);
   } catch (error) {
@@ -88,19 +80,30 @@ export async function DELETE(req) {
       return NextResponse.json({ error: "Not Authorized" }, { status: 401 });
     }
 
+    console.log("Board ID:", boardId);
+    console.log("Session User ID:", session?.user?.id);
+
     const user = await User.findById(session?.user?.id);
 
-
+    if (!user) {
+      return NextResponse.json({ error: "User not found" }, { status: 404 });
+    }
 
     await Board.deleteOne({
       _id: boardId,
       userId: session?.user?.id,
     });
 
+
     user.boards = user.boards.filter((id) => id.toString() !== boardId);
-    
-    return NextResponse.json({});
+    await user.save();
+
+    return NextResponse.json(
+      { message: "Board deleted successfully" },
+      { status: 200 }
+    );
   } catch (error) {
-    return NextResponse.json({ error: error.messsage }, { status: 500 });
+    console.error("Error deleting board:", error);
+    return NextResponse.json({ error: error.message }, { status: 500 });
   }
 }
